# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo Rest API",
        "description": "Este API permite implementar un sencillo sistema de 'To do list' con fines didácticos.",
        "contact": {
            "name": "Soporte API",
            "url": "https://escuela.it/contacto",
            "email": "contacto@escuela.it"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/mit/"
        },
        "version": "0.0"
    },
    "paths": {
        "/api/todos/{id}": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Show a specific todo",
                "operationId": "9891abbd1b15c9b858870c9fa96fcde9",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/acceptJsonHeader"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the requested todo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Todo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Todo not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Todos"
                ],
                "summary": "Update an existing todo",
                "operationId": "098f6dfbda208d015ac2494709fe1828",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/acceptJsonHeader"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the todo",
                                        "type": "string",
                                        "example": "Learn Laravel"
                                    },
                                    "description": {
                                        "description": "Description of the todo",
                                        "type": "string",
                                        "example": "Complete the Laravel tutorial"
                                    },
                                    "completed": {
                                        "description": "Completion status of the todo",
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Todo updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Todo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Todo not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Todos"
                ],
                "summary": "Delete a specific todo",
                "operationId": "ed4e0dc0d697bd72014fe9b032cdb406",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/acceptJsonHeader"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Todo deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Todo deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Todo not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/todos": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "List all todos",
                "operationId": "3341c7f75bd05712e7cf93f270ab9b02",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/acceptJsonHeader"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of todos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Todo"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Create a new todo",
                "operationId": "eb3d063eaea8969e943bd722408b0d3b",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/acceptJsonHeader"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the todo",
                                        "type": "string",
                                        "example": "Learn Laravel"
                                    },
                                    "description": {
                                        "description": "Description of the todo",
                                        "type": "string",
                                        "example": "Complete the Laravel tutorial"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Todo created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Todo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Todo": {
                "title": "Todo",
                "properties": {
                    "id": {
                        "description": "Todo ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Todo name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Todo description",
                        "type": "string"
                    },
                    "completed": {
                        "description": "Todo completion status",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "acceptJsonHeader": {
                "name": "Accept",
                "in": "header",
                "description": "Header to indicate the response format, should be application/json",
                "required": true,
                "schema": {
                    "type": "string",
                    "default": "application/json"
                }
            },
            "requestedWith": {
                "name": "X-Requested-With",
                "in": "header",
                "description": "Header to indicate the requested with parameter",
                "required": true,
                "schema": {
                    "type": "string",
                    "default": "XMLHttpRequest"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Todos",
            "description": "Todos"
        }
    ]
}